requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
yfinance==1.3.0
pandas==3.0.2
numpy==2.4.4
requests==2.33.1
apscheduler==3.11.2
python-dotenv==1.2.2
